feat: Input, Button, Table; refactor: change structure of the project
