+++ src/Button/Button.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+
+
+export class Button extends React.Component {
+    render() {
+        return (
+            <button>
+                { this.props.children }
+            </button>
+        )
+    }
+}
