+++ src/Table/Table.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+
+
+
+export class Table extends React.Component {
+    render() {
+        return 
+    }
+}
